requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# LLM and AI
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Embeddings and Vector Store
sentence-transformers>=2.3.0
faiss-cpu>=1.9.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Evaluation
# ragas>=0.1.0
# datasets>=2.0.0

# Utilities
tqdm>=4.66.0